query rest api/requirements.txt
click
request
python-dotenv